# Conflicts:
#	src/index/gpu_vector_index.cpp
#	src/index/multi_gpu_vector_index.cpp
#	src/llm/lora_framework/gpu_training_loop.cpp
#	src/llm/lora_framework/multi_gpu_trainer.cpp
#	src/llm/subagent_factory_impl.cpp
#	src/storage/columnar_format.cpp
#	src/updates/delta_update_engine.cpp

+++ src/auth/token_blacklist.cpp
@@ -55,7 +55,7 @@ void TokenBlacklist::revoke(const std::string &jti, std::chrono::system_clock::t
         }
 
         // Enforce hard cap
-        if (static_cast<int>(blacklist_.size()) >= config_.max_entries) {
+        if (blacklist_.size() >= static_cast<size_t>(config_.max_entries)) {
             THEMIS_WARN("TokenBlacklist: max_entries ({}) reached – dropping oldest entry", config_.max_entries);
             // Remove the entry that expires soonest (cheapest to lose)
             auto oldest = blacklist_.begin();
